Refactor service layer to split god object into focused services
- Split MeetingsService into separate services:
  - AuthService: JWT authentication and authorization
  - MeetingService: Meeting CRUD operations
  - MeetingRegistrantService: Registrant operations
  - PastMeetingService: Past meeting operations
  - PastMeetingParticipantService: Participant operations

- Move message/webhook handlers to internal/handlers package:
  - MeetingHandler: Handles NATS messages for meetings
  - ZoomWebhookHandler: Processes Zoom webhook events

- Update application layer (api.go) to use new service structure:
  - Inject individual services instead of god object
  - Update all endpoints to use appropriate service

- Refactor platform configuration:
  - Separate platform provider and webhook validator setup
  - Cleaner initialization in main.go

- Update CLAUDE.md and README.md to reflect new architecture

This improves separation of concerns, testability, and makes the codebase
more maintainable as it grows.
