+++ cmd/meeting-api/platforms/zoom.go
@@ -8,7 +8,6 @@ import (
 	"os"
 
 	"github.com/linuxfoundation/lfx-v2-meeting-service/internal/domain"
-	"github.com/linuxfoundation/lfx-v2-meeting-service/internal/domain/models"
 	"github.com/linuxfoundation/lfx-v2-meeting-service/internal/infrastructure/zoom"
 	"github.com/linuxfoundation/lfx-v2-meeting-service/internal/infrastructure/zoom/api"
 	"github.com/linuxfoundation/lfx-v2-meeting-service/internal/infrastructure/zoom/webhook"
@@ -20,6 +19,8 @@ type ZoomConfig struct {
 	ClientID           string
 	ClientSecret       string
 	WebhookSecretToken string
+	PlatformProvider   domain.PlatformProvider
+	Validator          domain.WebhookValidator
 }
 
 // NewZoomConfigFromEnv creates a ZoomConfig from environment variables
@@ -47,12 +48,12 @@ func (z ZoomConfig) ToAPIConfig() api.Config {
 }
 
 // SetupZoom configures Zoom platform integration and returns the webhook validator if configured
-func SetupZoom(platformRegistry domain.PlatformRegistry, config ZoomConfig) domain.WebhookValidator {
+func SetupZoom(config ZoomConfig) ZoomConfig {
 	// Setup Zoom platform provider
 	if config.IsConfigured() {
 		zoomClient := api.NewClient(config.ToAPIConfig())
 		zoomProvider := zoom.NewZoomProvider(zoomClient)
-		platformRegistry.RegisterProvider(models.PlatformZoom, zoomProvider)
+		config.PlatformProvider = zoomProvider
 
 		slog.Info("Zoom platform integration configured",
 			"account_id", config.AccountID,
@@ -68,9 +69,10 @@ func SetupZoom(platformRegistry domain.PlatformRegistry, config ZoomConfig) doma
 	if config.WebhookSecretToken != "" {
 		validator := webhook.NewZoomWebhookValidator(config.WebhookSecretToken)
 		slog.Info("Zoom webhook validation configured")
-		return validator
+		config.Validator = validator
+	} else {
+		slog.Warn("Zoom webhook validation not configured")
 	}
 
-	slog.Warn("Zoom webhook validation not configured - missing ZOOM_WEBHOOK_SECRET_TOKEN")
-	return nil
+	return config
 }
