+++ internal/handlers/meeting_handlers.go
@@ -1,7 +1,7 @@
 // Copyright The Linux Foundation and each contributor to LFX.
 // SPDX-License-Identifier: MIT
 
-package service
+package handlers
 
 import (
 	"context"
@@ -13,12 +13,42 @@ import (
 	"github.com/linuxfoundation/lfx-v2-meeting-service/internal/domain"
 	"github.com/linuxfoundation/lfx-v2-meeting-service/internal/domain/models"
 	"github.com/linuxfoundation/lfx-v2-meeting-service/internal/logging"
+	"github.com/linuxfoundation/lfx-v2-meeting-service/internal/service"
 	"github.com/linuxfoundation/lfx-v2-meeting-service/pkg/concurrent"
 	"github.com/linuxfoundation/lfx-v2-meeting-service/pkg/utils"
 )
 
+// ZoomWebhookHandler handles Zoom webhook events.
+type MeetingHandler struct {
+	meetingService                *service.MeetingService
+	registrantService             *service.MeetingRegistrantService
+	pastMeetingService            *service.PastMeetingService
+	pastMeetingParticipantService *service.PastMeetingParticipantService
+}
+
+func NewMeetingHandler(
+	meetingService *service.MeetingService,
+	registrantService *service.MeetingRegistrantService,
+	pastMeetingService *service.PastMeetingService,
+	pastMeetingParticipantService *service.PastMeetingParticipantService,
+) *MeetingHandler {
+	return &MeetingHandler{
+		meetingService:                meetingService,
+		registrantService:             registrantService,
+		pastMeetingService:            pastMeetingService,
+		pastMeetingParticipantService: pastMeetingParticipantService,
+	}
+}
+
+func (s *MeetingHandler) HandlerReady() bool {
+	return s.meetingService.ServiceReady() &&
+		s.registrantService.ServiceReady() &&
+		s.pastMeetingService.ServiceReady() &&
+		s.pastMeetingParticipantService.ServiceReady()
+}
+
 // HandleMessage implements domain.MessageHandler interface
-func (s *MeetingsService) HandleMessage(ctx context.Context, msg domain.Message) {
+func (s *MeetingHandler) HandleMessage(ctx context.Context, msg domain.Message) {
 	subject := msg.Subject()
 	ctx = logging.AppendCtx(ctx, slog.String("subject", subject))
 	slog.DebugContext(ctx, "handling NATS message")
@@ -27,16 +57,8 @@ func (s *MeetingsService) HandleMessage(ctx context.Context, msg domain.Message)
 	var err error
 
 	handlers := map[string]func(ctx context.Context, msg domain.Message) ([]byte, error){
-		models.MeetingGetTitleSubject:                         s.HandleMeetingGetTitle,
-		models.MeetingDeletedSubject:                          s.HandleMeetingDeleted,
-		models.ZoomWebhookMeetingStartedSubject:               s.HandleZoomMeetingStarted,
-		models.ZoomWebhookMeetingEndedSubject:                 s.HandleZoomMeetingEnded,
-		models.ZoomWebhookMeetingDeletedSubject:               s.HandleZoomMeetingDeleted,
-		models.ZoomWebhookMeetingParticipantJoinedSubject:     s.HandleZoomParticipantJoined,
-		models.ZoomWebhookMeetingParticipantLeftSubject:       s.HandleZoomParticipantLeft,
-		models.ZoomWebhookRecordingCompletedSubject:           s.HandleZoomRecordingCompleted,
-		models.ZoomWebhookRecordingTranscriptCompletedSubject: s.HandleZoomTranscriptCompleted,
-		models.ZoomWebhookMeetingSummaryCompletedSubject:      s.HandleZoomSummaryCompleted,
+		models.MeetingGetTitleSubject: s.HandleMeetingGetTitle,
+		models.MeetingDeletedSubject:  s.HandleMeetingDeleted,
 	}
 
 	handler, ok := handlers[subject]
@@ -77,9 +99,8 @@ func (s *MeetingsService) HandleMessage(ctx context.Context, msg domain.Message)
 	}
 }
 
-func (s *MeetingsService) handleMeetingGetAttribute(ctx context.Context, msg domain.Message, subject, getAttribute string) ([]byte, error) {
-
-	if !s.ServiceReady() {
+func (s *MeetingHandler) handleMeetingGetAttribute(ctx context.Context, msg domain.Message, subject, getAttribute string) ([]byte, error) {
+	if !s.meetingService.ServiceReady() {
 		slog.ErrorContext(ctx, "NATS KV store not initialized")
 		return nil, fmt.Errorf("NATS KV store not initialized")
 	}
@@ -96,7 +117,7 @@ func (s *MeetingsService) handleMeetingGetAttribute(ctx context.Context, msg dom
 		return nil, err
 	}
 
-	meeting, err := s.MeetingRepository.GetBase(ctx, meetingUID)
+	meeting, err := s.meetingService.MeetingRepository.GetBase(ctx, meetingUID)
 	if err != nil {
 		slog.ErrorContext(ctx, "error getting meeting from NATS KV", logging.ErrKey, err)
 		return nil, err
@@ -118,14 +139,14 @@ func (s *MeetingsService) handleMeetingGetAttribute(ctx context.Context, msg dom
 }
 
 // HandleMeetingGetTitle is the message handler for the meeting-get-title subject.
-func (s *MeetingsService) HandleMeetingGetTitle(ctx context.Context, msg domain.Message) ([]byte, error) {
+func (s *MeetingHandler) HandleMeetingGetTitle(ctx context.Context, msg domain.Message) ([]byte, error) {
 	return s.handleMeetingGetAttribute(ctx, msg, models.MeetingGetTitleSubject, "title")
 }
 
 // HandleMeetingDeleted is the message handler for the meeting-deleted subject.
 // It cleans up all registrants associated with the deleted meeting.
-func (s *MeetingsService) HandleMeetingDeleted(ctx context.Context, msg domain.Message) ([]byte, error) {
-	if !s.ServiceReady() {
+func (s *MeetingHandler) HandleMeetingDeleted(ctx context.Context, msg domain.Message) ([]byte, error) {
+	if !s.meetingService.ServiceReady() {
 		slog.ErrorContext(ctx, "service not ready")
 		return nil, fmt.Errorf("service not ready")
 	}
@@ -148,7 +169,7 @@ func (s *MeetingsService) HandleMeetingDeleted(ctx context.Context, msg domain.M
 	slog.InfoContext(ctx, "processing meeting deletion, cleaning up registrants")
 
 	// Get all registrants for the meeting
-	registrants, err := s.RegistrantRepository.ListByMeeting(ctx, meetingUID)
+	registrants, err := s.registrantService.RegistrantRepository.ListByMeeting(ctx, meetingUID)
 	if err != nil {
 		slog.ErrorContext(ctx, "error getting registrants for deleted meeting", logging.ErrKey, err)
 		return nil, err
@@ -167,7 +188,7 @@ func (s *MeetingsService) HandleMeetingDeleted(ctx context.Context, msg domain.M
 		reg := registrant // capture loop variable
 		tasks = append(tasks, func() error {
 			// Use the shared helper with skipRevisionCheck=true for bulk cleanup
-			err := s.deleteRegistrantWithCleanup(ctx, reg, 0, true)
+			err := s.registrantService.DeleteRegistrantWithCleanup(ctx, reg, 0, true)
 			if err != nil {
 				slog.ErrorContext(ctx, "error deleting registrant",
 					"registrant_uid", reg.UID,
