+++ internal/handlers/meeting_handlers_test.go
@@ -1,7 +1,7 @@
 // Copyright The Linux Foundation and each contributor to LFX.
 // SPDX-License-Identifier: MIT
 
-package service
+package handlers
 
 import (
 	"context"
